Restructure for Flow dist
Closes #159.
