--- src-node/.eslintrc.js
@@ -1,3 +1,0 @@
-module.exports = {
-  extends: '../conf/eslint-node.json'
-};
