--- src-webpack/.eslintrc.js
@@ -1,3 +1,0 @@
-module.exports = {
-  extends: '../conf/eslint-browser.json'
-};
