+++ src/node/.eslintrc.js
@@ -1,0 +1,3 @@
+module.exports = {
+  extends: '../../conf/eslint-node.json'
+};
