+++ src/node/render-pretty-error-stack.js
