+++ src/node/server-init-message.js
