+++ src/node/wrap-error.js
