+++ src/node/write-context-module.js
