+++ src/webpack/empty-application-wrapper.js
