+++ src/webpack/public/with-location.js
