+++ test/prefix-url.test.js
@@ -1,7 +1,7 @@
 'use strict';
 
-const prefixUrl = require('../src-webpack/public/prefix-url').prefixUrl;
-const prefixUrlAbsolute = require('../src-webpack/public/prefix-url')
+const prefixUrl = require('../src/webpack/public/prefix-url').prefixUrl;
+const prefixUrlAbsolute = require('../src/webpack/public/prefix-url')
   .prefixUrlAbsolute;
 
 describe('prefixUrl', () => {
